refactored css
