--- src/components/EventManagement/Dashboard/EventDashboard.module.css
@@ -1,101 +1,0 @@
-.eventContainer {
-  display: flex;
-  align-items: start;
-}
-
-.eventDetailsBox {
-  position: relative;
-  box-sizing: border-box;
-  background: #ffffff;
-  width: 66%;
-  padding: 0.3rem;
-  box-shadow: 0 3px 8px rgba(0, 0, 0, 0.1);
-  border-radius: 20px;
-  margin-bottom: 0;
-  margin-top: 20px;
-}
-.ctacards {
-  padding: 20px;
-  width: 100%;
-  display: flex;
-  background-color: #ffffff;
-  margin: 0 4px;
-  justify-content: space-between;
-  box-shadow: 0 3px 8px rgba(0, 0, 0, 0.1);
-  align-items: center;
-  border-radius: 20px;
-}
-.ctacards span {
-  color: rgb(181, 181, 181);
-  font-size: small;
-}
-/* .eventDetailsBox::before {
-  content: '';
-  position: absolute;
-  top: 0;
-  height: 100%;
-  width: 6px;
-  background-color: #31bb6b;
-  border-radius: 20px;
-} */
-
-.time {
-  display: flex;
-  justify-content: space-between;
-  padding: 15px;
-  padding-bottom: 0px;
-  width: 33%;
-
-  box-sizing: border-box;
-  background: #ffffff;
-  box-shadow: 0 3px 8px rgba(0, 0, 0, 0.1);
-  border-radius: 20px;
-  margin-bottom: 0;
-  margin-top: 20px;
-  margin-left: 10px;
-}
-
-.startTime,
-.endTime {
-  display: flex;
-  font-size: 20px;
-}
-
-.to {
-  padding-right: 10px;
-}
-
-.startDate,
-.endDate {
-  color: #808080;
-  font-size: 14px;
-}
-
-.titlename {
-  font-weight: 600;
-  font-size: 25px;
-  padding: 15px;
-  padding-bottom: 0px;
-  width: 50%;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-}
-
-.description {
-  color: #737373;
-  font-weight: 300;
-  font-size: 14px;
-  word-wrap: break-word;
-  padding: 15px;
-  padding-bottom: 0px;
-}
-
-.toporgloc {
-  font-size: 16px;
-  padding: 0.5rem;
-}
-
-.toporgloc span {
-  color: #737373;
-}
