--- src/components/EventManagement/EventAgendaItems/EventAgendaItems.module.css
@@ -1,22 +1,0 @@
-.eventAgendaItemContainer h2 {
-  margin: 0.6rem 0;
-}
-
-.btnsContainer {
-  display: flex;
-  gap: 10px;
-}
-
-@media (max-width: 768px) {
-  .btnsContainer {
-    margin-bottom: 0;
-    display: flex;
-    flex-direction: column;
-  }
-
-  .createAgendaItemButton {
-    position: absolute;
-    top: 1rem;
-    right: 2rem;
-  }
-}
