--- src/components/EventManagement/EventAttendance/EventsAttendance.module.css
@@ -1,35 +1,0 @@
-.input {
-  display: flex;
-  width: 100%;
-  position: relative;
-}
-.customcell {
-  background-color: #31bb6b !important;
-  color: white !important;
-  font-size: medium !important;
-  font-weight: 500 !important;
-  padding-top: 10px !important;
-  padding-bottom: 10px !important;
-}
-
-.eventsAttended,
-.membername {
-  color: blue;
-}
-.actionBtn {
-  /* color:#39a440 !important; */
-  background-color: #ffffff !important;
-}
-.actionBtn:hover,
-.actionBtn:focus,
-.actionBtn:active {
-  color: #39a440 !important;
-}
-
-.table-body > .table-row {
-  background-color: #fff !important;
-}
-
-.table-body > .table-row:nth-child(2n) {
-  background: #afffe8 !important;
-}
