--- src/screens/OrganizationActionItems/OrganizationActionItems.module.css
@@ -1,291 +1,0 @@
-actionItemsContainer {
-  height: 90vh;
-}
-
-.actionItemModal {
-  max-width: 80vw;
-  margin-top: 2vh;
-  margin-left: 13vw;
-}
-
-.datediv {
-  display: flex;
-  flex-direction: row;
-}
-
-.datebox {
-  width: 90%;
-  border-radius: 7px;
-  outline: none;
-  box-shadow: none;
-  padding-top: 2px;
-  padding-bottom: 2px;
-  padding-right: 5px;
-  padding-left: 5px;
-  margin-right: 5px;
-  margin-left: 5px;
-}
-
-.dropdownToggle {
-  margin-bottom: 0;
-  display: flex;
-}
-
-.dropdownModalToggle {
-  width: 50%;
-}
-
-.errorIcon {
-  transform: scale(1.5);
-  color: var(--bs-danger);
-  margin-bottom: 1rem;
-}
-
-.greenregbtn {
-  margin: 1rem 0 0;
-  margin-top: 15px;
-  border: 1px solid var(--bs-gray-300);
-  box-shadow: 0 2px 2px var(--bs-gray-300);
-  padding: 10px 10px;
-  border-radius: 5px;
-  background-color: var(--bs-primary);
-  width: 100%;
-  font-size: 16px;
-  color: var(--bs-white);
-  outline: none;
-  font-weight: 600;
-  cursor: pointer;
-  transition:
-    transform 0.2s,
-    box-shadow 0.2s;
-  width: 100%;
-}
-
-hr {
-  border: none;
-  height: 1px;
-  background-color: var(--bs-gray-500);
-  margin: 1rem;
-}
-
-.iconContainer {
-  display: flex;
-  justify-content: flex-end;
-}
-.icon {
-  margin: 1px;
-}
-
-.message {
-  margin-top: 25%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.preview {
-  display: flex;
-  flex-direction: row;
-  font-weight: 900;
-  font-size: 16px;
-  color: rgb(80, 80, 80);
-}
-
-.removeFilterIcon {
-  cursor: pointer;
-}
-
-.searchForm {
-  display: inline;
-}
-
-.view {
-  margin-left: 2%;
-  font-weight: 600;
-  font-size: 16px;
-  color: var(--bs-gray-600);
-}
-
-/* header (search, filter, dropdown) */
-.btnsContainer {
-  display: flex;
-  margin: 0.5rem 0 1.5rem 0;
-}
-
-.btnsContainer .input {
-  flex: 1;
-  min-width: 18rem;
-  position: relative;
-}
-
-.btnsContainer input {
-  outline: 1px solid var(--bs-gray-400);
-}
-
-.btnsContainer .input button {
-  width: 52px;
-}
-
-.noOutline input {
-  outline: none;
-}
-
-.noOutline input:disabled {
-  -webkit-text-fill-color: black !important;
-}
-
-.noOutline textarea:disabled {
-  -webkit-text-fill-color: black !important;
-}
-
-.inputField {
-  margin-top: 10px;
-  margin-bottom: 10px;
-  background-color: white;
-  box-shadow: 0 1px 1px #31bb6b;
-}
-
-.inputField > button {
-  padding-top: 10px;
-  padding-bottom: 10px;
-}
-
-.dropdown {
-  background-color: white;
-  border: 1px solid #31bb6b;
-  position: relative;
-  display: inline-block;
-  color: #31bb6b;
-}
-
-/* Action Items Data Grid */
-.rowBackground {
-  background-color: var(--bs-white);
-  max-height: 120px;
-}
-
-.tableHeader {
-  background-color: var(--bs-primary);
-  color: var(--bs-white);
-  font-size: 1rem;
-}
-
-.chipIcon {
-  height: 0.9rem !important;
-}
-
-.chip {
-  height: 1.5rem !important;
-}
-
-.active {
-  background-color: #31bb6a50 !important;
-}
-
-.pending {
-  background-color: #ffd76950 !important;
-  color: #bb952bd0 !important;
-  border-color: #bb952bd0 !important;
-}
-
-/* Modals */
-.itemModal {
-  max-width: 80vw;
-  margin-top: 2vh;
-  margin-left: 13vw;
-}
-
-.titlemodal {
-  color: #707070;
-  font-weight: 600;
-  font-size: 32px;
-  width: 65%;
-  margin-bottom: 0px;
-}
-
-.modalCloseBtn {
-  width: 40px;
-  height: 40px;
-  padding: 1rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.imageContainer {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  margin-right: 0.5rem;
-}
-
-.TableImage {
-  object-fit: cover;
-  width: 25px !important;
-  height: 25px !important;
-  border-radius: 100% !important;
-}
-
-.avatarContainer {
-  width: 28px;
-  height: 26px;
-}
-
-/* Toggle Btn */
-.toggleGroup {
-  width: 50%;
-  min-width: 20rem;
-  margin: 0.5rem 0rem;
-}
-
-.toggleBtn {
-  padding: 0rem;
-  height: 2rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.toggleBtn:hover {
-  color: #31bb6b !important;
-}
-
-input[type='radio']:checked + label {
-  background-color: #31bb6a50 !important;
-}
-
-input[type='radio']:checked + label:hover {
-  color: black !important;
-} /* Toggle Btn */
-.toggleGroup {
-  width: 50%;
-  min-width: 20rem;
-  margin: 0.5rem 0rem;
-}
-
-.toggleBtn {
-  padding: 0rem;
-  height: 2rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.toggleBtn:hover {
-  color: #31bb6b !important;
-}
-
-input[type='radio']:checked + label {
-  background-color: #31bb6a50 !important;
-}
-
-input[type='radio']:checked + label:hover {
-  color: black !important;
-}
-
-.rankings {
-  aspect-ratio: 1;
-  border-radius: 50%;
-  width: 50px;
-}
